move helper function to theming helpers
